+++ lib/porous/cli.rb
@@ -12,5 +12,6 @@ require 'rack'
 require 'rackup/server'
 
 require 'porous/cli/build'
+require 'porous/cli/dev'
 require 'porous/cli/new'
 require 'porous/cli/server'