+++ lib/porous/cli/server.rb
@@ -6,19 +6,7 @@ module Porous
 
     namespace :server
 
-    desc 'server [OPTIONS]', 'Starts Porous server'
-    method_option :port,
-                  aliases: :p,
-                  type: :numeric,
-                  default: 9292,
-                  desc: 'The port Porous will listen on'
-
-    method_option :host,
-                  aliases: :h,
-                  type: :string,
-                  default: 'localhost',
-                  desc: 'The host address Porous will bind to'
-
+    desc 'server [OPTIONS]', 'Starts Porous server in production mode'
     def server # rubocop:todo Metrics/MethodLength
       Agoo::Log.configure(dir: '',
                           console: true,
@@ -27,26 +15,33 @@ module Porous
                           states: {
                             INFO: true,
                             DEBUG: false,
-                            connect: false,
-                            request: false,
+                            connect: true,
+                            request: true,
                             response: false,
-                            eval: true,
-                            push: true
+                            eval: false,
+                            push: false
                           })
 
-      Agoo::Server.init 9292, Dir.pwd, thread_count: 1
+      Agoo::Server.init(
+        80, '.',
+        thread_count: 0,
+        ssl_cert: 'ssl/cert.pem',
+        ssl_key: 'ssl/key.pem',
+        bind: [
+          'http://127.0.0.1:80',
+          'https://127.0.0.1:443'
+        ]
+      )
       Agoo::Server.use Rack::ContentLength
       Agoo::Server.use Rack::Static, urls: ['/static']
-      Agoo::Server.use Rack::ShowExceptions
 
       # Socket Communication
       $socket ||= Porous::Server::Socket.new
       Agoo::Server.handle nil, '/connect', Porous::Server::Connect.new
       # Server-Side Rendering
       Agoo::Server.handle nil, '**', Porous::Server::Application.new
+
       Agoo::Server.start
-      # Live Reload Builder
-      Server::Builder.new.build.start
     end
   end
 end